restructure adapters
- make adapters name more infomative
  - Mock and Real -> ReadOnly and ReadAndWrite
- move methods that uses Yao from Resource to adapters
