+++ lib/kaname.rb
@@ -9,5 +9,3 @@ require 'kaname/resource'
 
 module Kaname
 end
-
-Kaname::Config.setup unless ENV['RACK_ENV'] == 'test'
